Separate loaders from scene.
