+++ src/scene.js
@@ -1,8 +1,5 @@
 const THREE = require('three');
-const resolveURL = require('./utilities').resolveURL;
-const JSONLoader = require('./loader').JSONLoader;
-const STLLoader = require('./STLLoader').STLLoader;
-const OBJLoader = require('./OBJLoader').OBJLoader;
+const SceneLoader =require('./sceneLoader').SceneLoader;
 
 /**
  * A Zinc.Scene contains {@link Zinc.Geometry}, {@link Zinc.Glyphset} and 
@@ -22,6 +19,7 @@ exports.Scene = function(containerIn, rendererIn) {
   let zincPointsets = [];
   let zincLines = [];
   let videoHandler = undefined;
+  let sceneLoader = new SceneLoader(this);
   const scene = new THREE.Scene();
   /**
    * A {@link THREE.DirectionalLight} object for controlling lighting of this scene.
@@ -38,8 +36,6 @@ exports.Scene = function(containerIn, rendererIn) {
   let num_inputs = 0;
   let startingId = 1000;
   this.sceneName = undefined;
-  this.progressMap = [];
-  let errorDownload = false;
   let stereoEffectFlag = false;
   let stereoEffect = undefined;
   this.autoClearFlag = true;
@@ -69,36 +65,8 @@ exports.Scene = function(containerIn, rendererIn) {
    * @returns {Array} 
    */
   this.getDownloadProgress = () => {
-    let totalSize = 0;
-    let totalLoaded = 0;
-    let unknownFound = false;
-
-    for (const key in this.progressMap) {
-      const progress = this.progressMap[key];
-
-      totalSize += progress[1];
-      totalLoaded += progress[0];
-
-      if (progress[1] == 0)
-        unknownFound = true;
-    }
-    if (unknownFound) {
-      totalSize = 0;
-    }
-    return [ totalSize, totalLoaded, errorDownload ];
+    return sceneLoader.getDownloadProgress();
   }
-
-  //Stores the current progress of downloads
-  this.onProgress = id => {
-    return xhr => {
-      this.progressMap[id] = [ xhr.loaded, xhr.total ];
-    };
-  }
-
-  this.onError = xhr => {
-	  num_inputs = num_inputs - 1;
-	  errorDownload = true;
-  };
 
   //called from Renderer when panel has been resized
   this.onWindowResize = () => {
@@ -130,21 +98,6 @@ exports.Scene = function(containerIn, rendererIn) {
   };
 
   setupCamera();
-
-  //Get the next available unique identifier for Zinc.Geometry
-  const nextAvailableInternalZincModelId = () => {
-    let idFound = true;
-    while (idFound == true) {
-      startingId++;
-      idFound = false
-      for (let i = 0; i < zincGeometries.length; i++) {
-        if (zincGeometries[i].modelId == startingId) {
-          idFound = true;
-        }
-      }
-    }
-    return startingId;
-  };
 
   /**
    * Load the viewport Data from the argument  {@link Zinc.Viewport} and set it as 
@@ -352,24 +305,19 @@ exports.Scene = function(containerIn, rendererIn) {
 	  }  
   }
 
-  //Load a glyphset into this scene.
-  const loadGlyphset = (glyphsetData, glyphurl, groupName, finishCallback) => {
-    const newGlyphset = new (require('./glyphset').Glyphset)();
-    newGlyphset.duration = 3000;
-    newGlyphset.load(glyphsetData, resolveURL(glyphurl), finishCallback);
-    newGlyphset.groupName = groupName;
-    this.addGlyphset(newGlyphset);
-  };
+  this.addLines = lines => {
+    if (lines && lines.isLines) {
+      scene.add(lines.morph);
+      zincLines.push(lines);
+    }
+  }
 
-  //Load a glyphset into this scene.
-  const onLoadGlyphsetReady = (xmlhttp, glyphurl, groupName, finishCallback) => {
-    return () => {
-      if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-        const glyphsetData = JSON.parse(xmlhttp.responseText);
-        loadGlyphset(glyphsetData, glyphurl, groupName, finishCallback);
-      }
-    };
-  };
+  this.addPointset = pointset => {
+    if (pointset && pointset.isPointset) {
+      scene.add(pointset.morph);
+      zincPointsets.push(pointset);
+    }
+  }
 
   /**
    * Load a glyphset into this scene object.
@@ -382,69 +330,8 @@ exports.Scene = function(containerIn, rendererIn) {
    * once the glyphset is succssfully load in.
    */
   this.loadGlyphsetURL = (metaurl, glyphurl, groupName, finishCallback) => {
-    const xmlhttp = new XMLHttpRequest();
-    xmlhttp.onreadystatechange = onLoadGlyphsetReady(xmlhttp, glyphurl, groupName, finishCallback);
-    xmlhttp.open("GET", resolveURL(metaurl), true);
-    xmlhttp.send();
+    sceneLoader.loadGlyphsetURL(metaurl, glyphurl, groupName, finishCallback);
   }
-  
-  //Internal loader for a regular zinc geometry.
-  const pointsetloader = (localTimeEnabled, localMorphColour, groupName, finishCallback) => {
-	  return (geometry, materials) => {
-		  const newPointset = new (require('./pointset').Pointset)();
-		  let material = new THREE.PointsMaterial({alphaTest:0.5,size: 5, sizeAttenuation:false});
-		  if (materials && materials[0]) {
-		      if (1.0 > materials[0].opacity) {
-		    	  material.transparent = true;
-		      }
-		      material.opacity = materials[0].opacity;
-			  material.color = materials[0].color;
-			  material.morphTargets = localTimeEnabled;
-			  material.vertexColors = materials[0].vertexColors;
-      }
-      let options = {};
-      options.localTimeEnabled = localTimeEnabled;
-      options.localMorphColour = localMorphColour;
-
-		  newPointset.createMesh(geometry, material, options);
-		  if (newPointset) {
-			  newPointset.setName(groupName);
-			  scene.add(newPointset.morph);
-			  zincPointsets.push(newPointset);
-		  }
-		  if (finishCallback != undefined && (typeof finishCallback == 'function'))
-			  finishCallback(newPointset);
-	  };
-  }
-
-  //Internal loader for a regular zinc geometry.
-  const linesloader = (localTimeEnabled, localMorphColour, groupName, finishCallback) => {
-    return (geometry, materials) => {
-      const newLines = new (require('./lines').Lines)();
-      let material = undefined;
-      if (materials && materials[0]) {
-        material = new THREE.LineBasicMaterial({color:materials[0].color.clone()});
-        if (1.0 > materials[0].opacity) {
-          material.transparent = true;
-        }
-        material.opacity = materials[0].opacity;
-        material.morphTargets = localTimeEnabled;
-        material.vertexColors = materials[0].vertexColors;
-      }
-      let options = {};
-      options.localTimeEnabled = localTimeEnabled;
-      options.localMorphColour = localMorphColour;
-
-      newLines.createLineSegment(geometry, material, options);
-      if (newLines) {
-        newLines.setName(groupName);
-        scene.add(newLines.morph);
-        zincLines.push(newLines);
-      }
-      if (finishCallback != undefined && (typeof finishCallback == 'function'))
-        finishCallback(newLines);
-    };
-  } 
 
   /**
    * Load a pointset into this scene object.
@@ -458,17 +345,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * once the glyphset is succssfully load in.
    */
   this.loadPointsetURL = (url, timeEnabled, morphColour, groupName, finishCallback) => {
-	  let localTimeEnabled = 0;
-	  num_inputs += 1;
-	  if (timeEnabled != undefined)
-		  localTimeEnabled = timeEnabled ? true : false;
-	  let localMorphColour = 0;
-	  if (morphColour != undefined)
-		  localMorphColour = morphColour ? true : false;
-	  let loader = new JSONLoader();
-	  loader.crossOrigin = "Anonymous";
-	  loader.load(url, pointsetloader(localTimeEnabled, localMorphColour, groupName, finishCallback),
-		  this.onProgress(i), this.onError);
+    sceneLoader.loadPointsetURL(url, timeEnabled, morphColour, groupName, finishCallback);
   }
 
     /**
@@ -483,100 +360,10 @@ exports.Scene = function(containerIn, rendererIn) {
    * once the glyphset is succssfully load in.
    */
   this.loadLinesURL = (url, timeEnabled, morphColour, groupName, finishCallback) => {
-	  let localTimeEnabled = 0;
-	  num_inputs += 1;
-	  if (timeEnabled != undefined)
-		  localTimeEnabled = timeEnabled ? true : false;
-	  let localMorphColour = 0;
-	  if (morphColour != undefined)
-		  localMorphColour = morphColour ? true : false;
-	  let loader = new JSONLoader();
-	  loader.crossOrigin = "Anonymous";
-	  loader.load(url, linesloader(localTimeEnabled, localMorphColour, groupName, finishCallback),
-		  this.onProgress(i), this.onError);
+    sceneLoader.loadPointsetURL(url, timeEnabled, morphColour, groupName, finishCallback);
   }
 
-  /**
-   * Load a geometry into this scene, this is a subsequent called from 
-   * {@link Zinc.Scene#loadMetadataURL}, although it can be used to read
-   * in geometry into the scene externally.
-   * 
-   * @param {String} url - regular json model file providing geometry.
-   * @param {Boolean} timeEnabled - Indicate if geometry morphing is enabled.
-   * @param {Boolean} morphColour - Indicate if color morphing is enabled.
-   * @param {STRING} groupName - name to assign the geometry's groupname to.
-   * @param {STRING} fileFormat - name supported formats are STL, OBJ and JSON.
-   * @param {Function} finishCallback - Callback function which will be called
-   * once the geometry is succssfully loaded in.
-   */
-  const loadMetaModel = (url, timeEnabled, morphColour, groupName, fileFormat, finishCallback) => {
-    num_inputs += 1;
-    const modelId = nextAvailableInternalZincModelId();
-    const colour = require('./zinc').defaultMaterialColor;
-    const opacity = require('./zinc').defaultOpacity;
-    let localTimeEnabled = 0;
-    if (timeEnabled != undefined)
-      localTimeEnabled = timeEnabled ? true : false;
-    let localMorphColour = 0;
-    if (morphColour != undefined)
-      localMorphColour = morphColour ? true : false;
-    let loader = new JSONLoader();
-    if (fileFormat !== undefined) {
-      if (fileFormat == "STL") {
-        loader = new STLLoader();
-      } else if (fileFormat == "OBJ") {
-        loader = new OBJLoader();
-  	    loader.crossOrigin = "Anonymous";
-        loader.load(url, objloader(modelId, colour, opacity, localTimeEnabled,
-          localMorphColour, groupName, finishCallback), this.onProgress(i), this.onError);
-        return;
-      }
-    }
-    loader.crossOrigin = "Anonymous";
-    loader.load(url, meshloader(modelId, colour, opacity, localTimeEnabled,
-      localMorphColour, groupName, finishCallback), this.onProgress(i), this.onError);
-  };
-  
-  //Object to keep track of number of items downloaded and when add items are downloaded
-  //allCompletedCallback is called
-  var metaFinishCallback = function(numberOfDownloaded, finishCallback, allCompletedCallback) {
-    let downloadedItem = 0;
-    return zincGeometry => {
-      downloadedItem = downloadedItem + 1;
-      if (finishCallback != undefined && (typeof finishCallback == 'function'))
-        finishCallback(zincGeometry);
-      if (downloadedItem == numberOfDownloaded)
-        if (allCompletedCallback != undefined && (typeof allCompletedCallback == 'function'))
-          allCompletedCallback();
-    };
-  };
-
-  //Function to process each of the metadata item. There are two types of metadata item,
-  //one for Zinc.Geometry and one for Zinc.Glyphset.
-  const readMetadataItem = (referenceURL, item, finishCallback) => {
-    if (item) {
-    	let newURL = item.URL;
-    	if (referenceURL)
-    	   	newURL = (new URL(item.URL, referenceURL)).href;
-      if (item.Type == "Surfaces") {
-        loadMetaModel(newURL, item.MorphVertices, item.MorphColours, item.GroupName, item.FileFormat, finishCallback);
-      } else if (item.Type == "Glyph") {
-      	let newGeometryURL = item.GlyphGeometriesURL;
-    	if (referenceURL) {
-    	   	newGeometryURL = (new URL(item.GlyphGeometriesURL, referenceURL)).href;
-    	}
-        this.loadGlyphsetURL(newURL, newGeometryURL, item.GroupName, finishCallback);
-      } else if (item.Type == "Points") {
-    	  this.loadPointsetURL(newURL, item.MorphVertices, item.MorphColours, item.GroupName, finishCallback);
-      } else if (item.Type == "Lines") {
-    	  this.loadLinesURL(newURL, item.MorphVertices, item.MorphColours, item.GroupName, finishCallback);
-      }else if (item.Type == "View") {
-    	  this.loadViewURL(newURL);
-      }
-    }
-  };
-
-  /**
+   /**
    * Read a STL file into this scene, the geometry will be presented as
    * {@link Zinc.Geometry}. 
    * 
@@ -586,14 +373,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * once the STL geometry is succssfully loaded.
    */
   this.loadSTL = (url, groupName, finishCallback) => {
-    num_inputs += 1;
-    const modelId = nextAvailableInternalZincModelId();
-    const colour = require('./zinc').defaultMaterialColor;
-    const opacity = require('./zinc').defaultOpacity;
-    const loader = new STLLoader();
-    loader.crossOrigin = "Anonymous";
-    loader.load(resolveURL(url), meshloader(modelId, colour, opacity, false,
-      false, groupName, finishCallback));
+    sceneLoader.loadSTL(url, groupName, finishCallback);
   }
 
   /**
@@ -606,14 +386,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * once the OBJ geometry is succssfully loaded.
    */
   this.loadOBJ = (url, groupName, finishCallback) => {
-    num_inputs += 1;
-    const modelId = nextAvailableInternalZincModelId();
-    const colour = require('./zinc').defaultMaterialColor;
-    const opacity = require('./zinc').defaultOpacity;
-    const loader = new OBJLoader();
-    loader.crossOrigin = "Anonymous";
-    loader.load(resolveURL(url), meshloader(modelId, colour, opacity, false,
-      false, groupName, finishCallback));
+    sceneLoader.loadOBJ(url, groupName, finishCallback);
   }
 
   /**
@@ -625,28 +398,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * for each glyphset and geometry that has been written in.
    */
   this.loadMetadataURL = (url, finishCallback, allCompletedCallback) => {
-    const xmlhttp = new XMLHttpRequest();
-    var requestURL = resolveURL(url);
-    xmlhttp.onreadystatechange = () => {
-      if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-    	let  referenceURL = xmlhttp.responseURL;
-    	if (referenceURL === undefined)
-    		referenceURL = (new URL(requestURL)).href;
-        const metadata = JSON.parse(xmlhttp.responseText);
-        let numberOfObjects = metadata.length;
-        // view file does not receive callback
-        for (var i = 0; i < metadata.length; i++) {
-        	if (metadata[i] && metadata[i].Type == "View")
-        		numberOfObjects = numberOfObjects - 1;
-        }   
-        var callback = new metaFinishCallback(numberOfObjects, finishCallback, allCompletedCallback);
-        for (var i = 0; i < metadata.length; i++)
-          readMetadataItem(referenceURL, metadata[i], callback);
-      }
-    }
- 
-    xmlhttp.open("GET", requestURL, true);
-    xmlhttp.send();
+    sceneLoader.loadMetadataURL(url, finishCallback, allCompletedCallback);
   }
 
   /**
@@ -656,28 +408,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * @deprecated
    */
   this.loadModelsURL = (urls, colours, opacities, timeEnabled, morphColour, finishCallback) => {
-    const number = urls.length;
-    num_inputs += number;
-    for (let i = 0; i < number; i++) {
-      const modelId = nextAvailableInternalZincModelId();
-      const filename = urls[i];
-      const loader = new JSONLoader();
-      let colour = require('./zinc').defaultMaterialColor;
-      let opacity = require('./zinc').defaultOpacity;
-      if (colours != undefined && colours[i] != undefined)
-        colour = colours[i] ? true : false;
-      if (opacities != undefined && opacities[i] != undefined)
-        opacity = opacities[i];
-      let localTimeEnabled = 0;
-      if (timeEnabled != undefined && timeEnabled[i] != undefined)
-        localTimeEnabled = timeEnabled[i] ? true : false;
-      let localMorphColour = 0;
-      if (morphColour != undefined && morphColour[i] != undefined)
-        localMorphColour = morphColour[i] ? true : false;
-      loader.crossOrigin = "Anonymous";
-      loader.load(resolveURL(filename), meshloader(modelId, colour, opacity, localTimeEnabled, localMorphColour, undefined,
-        finishCallback), this.onProgress(i), this.onError);
-    }
+    sceneLoader.loadModelsURL(urls, colours, opacities, timeEnabled, morphColour, finishCallback);
   }
 
   /**
@@ -685,16 +416,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * @param {String} URL - address to the file containing viewport information.
    */
   this.loadViewURL = url => {
-    const xmlhttp = new XMLHttpRequest();
-    xmlhttp.onreadystatechange = () => {
-      if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-        const viewData = JSON.parse(xmlhttp.responseText);
-        this.loadView(viewData);
-      }
-    }
-    requestURL = resolveURL(url);
-    xmlhttp.open("GET", requestURL, true);
-    xmlhttp.send();
+    sceneLoader.loadViewURL(url);
   }
 
   /**
@@ -706,23 +428,7 @@ exports.Scene = function(containerIn, rendererIn) {
    * @deprecated
    */
   this.loadFromViewURL = (jsonFilePrefix, finishCallback) => {
-    const xmlhttp = new XMLHttpRequest();
-    xmlhttp.onreadystatechange = () => {
-      if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-        const viewData = JSON.parse(xmlhttp.responseText);
-        this.loadView(viewData);
-        const urls = [];
-        const filename_prefix = jsonFilePrefix + "_";
-        for (let i = 0; i < viewData.numberOfResources; i++) {
-          const filename = filename_prefix + (i + 1) + ".json";
-          urls.push(filename);
-        }
-        this.loadModelsURL(urls, viewData.colour, viewData.opacity, viewData.timeEnabled, viewData.morphColour, finishCallback);
-      }
-    }
-    requestURL = resolveURL(jsonFilePrefix + "_view.json");
-    xmlhttp.open("GET", requestURL, true);
-    xmlhttp.send();
+    sceneLoader.loadFromViewURL(jsonFilePrefix, finishCallback);
   }
 
   const setPositionOfObject = mesh => {
@@ -738,8 +444,8 @@ exports.Scene = function(containerIn, rendererIn) {
   this.addGeometry = zincGeometry => {
     if (zincGeometry && zincGeometry.morph) {
       if (zincGeometry.modelId === -1) {
-        num_inputs++;
-        zincGeometry.modelId = nextAvailableInternalZincModelId();
+        sceneLoader.num_inputs++;
+        zincGeometry.modelId = sceneLoader.nextAvailableInternalZincModelId();
       }
       scene.add(zincGeometry.morph);
       zincGeometries.push(zincGeometry) ;
@@ -753,31 +459,7 @@ exports.Scene = function(containerIn, rendererIn) {
     return newGeometry;
   };
 
-  //Loader for the OBJ format, 
-  const objloader = (
-    modelId,
-    colour,
-    opacity,
-    localTimeEnabled,
-    localMorphColour,
-    groupName,
-    finishCallback
-  ) => {
-    return object => {
-      num_inputs++;
-      object.traverse(child => {
-        if (child instanceof THREE.Mesh) {
-          const zincGeometry = addMeshToZincGeometry(child, modelId, localTimeEnabled, localMorphColour);
-          this.addGeometry(zincGeometry);
-          if (zincGeometry.morph)
-            zincGeometry.morph.name = groupName;
-          zincGeometry.groupName = groupName;
-          if (finishCallback != undefined && (typeof finishCallback == 'function'))
-            finishCallback(zincGeometry);
-        }
-      });
-    };
-  }
+  
   
   /**
    * Add a user provided {THREE.Geometry} into  the scene as zinc geometry.
@@ -820,7 +502,7 @@ exports.Scene = function(containerIn, rendererIn) {
       if (external == undefined)
         external = true
       if (external)
-        num_inputs++;
+        sceneLoader.num_inputs++;
       if (finishCallback != undefined && (typeof finishCallback == 'function'))
         finishCallback(newGeometry);
       if (!videoHandler && newGeometry.videoHandler)
@@ -828,27 +510,6 @@ exports.Scene = function(containerIn, rendererIn) {
       return newGeometry;
     }
     return undefined;
-  }
-
-  //Internal loader for a regular zinc geometry.
-  const meshloader = (
-    modelId,
-    colour,
-    opacity,
-    localTimeEnabled,
-    localMorphColour,
-    groupName,
-    finishCallback
-  ) => {
-    return (geometry, materials) => {
-      let material = undefined;
-      if (materials && materials[0]) {
-        material = materials[0];
-      }
-      const zincGeometry = this.addZincGeometry(geometry, modelId, colour, opacity, localTimeEnabled, localMorphColour, false, undefined, material, groupName);
-      if (finishCallback != undefined && (typeof finishCallback == 'function'))
-        finishCallback(zincGeometry);
-    };
   }
 
   //Update the directional light for this scene.
@@ -992,7 +653,7 @@ exports.Scene = function(containerIn, rendererIn) {
 			  }
 			  var currentTime = videoHandler.video.currentTime / videoHandler.getVideoDuration() * 3000;
 			  var totalInput = zincGeometries.length + zincPointsets.length;
-			  if (totalInput == num_inputs && allGlyphsetsReady()) {
+			  if (totalInput == sceneLoader.num_inputs && allGlyphsetsReady()) {
 				  zincCameraControls.setTime(currentTime);
 				  zincCameraControls.update(0);
 				  for (let i = 0; i < zincGeometries.length; i++) {
@@ -1021,7 +682,7 @@ exports.Scene = function(containerIn, rendererIn) {
 		  }
 	  } else {
 		  var totalInput = zincGeometries.length + zincPointsets.length + zincLines.length;
-		  if (totalInput == num_inputs && allGlyphsetsReady()) {
+		  if (totalInput == sceneLoader.num_inputs && allGlyphsetsReady()) {
 			  zincCameraControls.update(delta);
 			  for (let i = 0; i < zincGeometries.length; i++) {
 				  /* check if morphColour flag is set */
